+++ src/app/dim-api/reducer.ts
@@ -419,7 +419,7 @@ export const dimApi = (
 
     case getType(actions.searchDeleted):
       return produce(state, (draft) => {
-        deleteSearch(draft, account!.destinyVersion, action.payload.query, action.payload.type);
+        deleteSearch(account!, draft, action.payload.query, action.payload.type);
       });
 
     // *** Triumphs ***
@@ -463,23 +463,20 @@ function migrateSettings(state: DimApiState) {
   }
 
   // Replace 'element' sort with 'elementWeapon' and 'elementArmor'
-  const sortOrder = state.settings.itemSortOrderCustom || [];
-  const reversals = state.settings.itemSortReversals || [];
+  const sortOrder = [...new Set(state.settings.itemSortOrderCustom || [])];
+  const reversals = [...new Set(state.settings.itemSortReversals || [])];
 
   if (sortOrder.includes('element')) {
-    state = changeSetting(
-      state,
-      'itemSortOrderCustom',
-      sortOrder.toSpliced(sortOrder.indexOf('element'), 1, 'elementWeapon', 'elementArmor'),
-    );
+    sortOrder.splice(sortOrder.indexOf('element'), 1, 'elementWeapon', 'elementArmor');
   }
-
+  if (sortOrder.length !== (state.settings.itemSortOrderCustom?.length ?? 0)) {
+    state = changeSetting(state, 'itemSortOrderCustom', sortOrder);
+  }
   if (reversals.includes('element')) {
-    state = changeSetting(
-      state,
-      'itemSortReversals',
-      reversals.toSpliced(sortOrder.indexOf('element'), 1, 'elementWeapon', 'elementArmor'),
-    );
+    reversals.splice(reversals.indexOf('element'), 1, 'elementWeapon', 'elementArmor');
+  }
+  if (reversals.length !== (state.settings.itemSortReversals?.length ?? 0)) {
+    state = changeSetting(state, 'itemSortReversals', reversals);
   }
 
   // converts any old custom stats stored in the old settings key, to the new format
@@ -1145,7 +1142,8 @@ function searchUsed(
       query,
       type,
     },
-    destinyVersion,
+    platformMembershipId: account.membershipId,
+    destinyVersion: account.destinyVersion,
   };
   applyUpdateLocally(draft, updateAction);
   draft.updateQueue.push(updateAction);
@@ -1164,7 +1162,7 @@ function searchUsed(
       const lastSearch = sortedSearches.pop()!;
       // Never try to delete the built-in searches or saved searches
       if (!lastSearch.saved && lastSearch.usageCount > 0) {
-        deleteSearch(draft, destinyVersion, lastSearch.query, lastSearch.type);
+        deleteSearch(account, draft, lastSearch.query, lastSearch.type);
       }
     }
   }
@@ -1205,7 +1203,8 @@ function saveSearch(
         query,
         type,
       },
-      destinyVersion,
+      platformMembershipId: account.membershipId,
+      destinyVersion: account.destinyVersion,
     };
     applyUpdateLocally(draft, searchUsedUpdate);
     draft.updateQueue.push(searchUsedUpdate);
@@ -1218,15 +1217,16 @@ function saveSearch(
       saved,
       type,
     },
-    destinyVersion,
+    platformMembershipId: account.membershipId,
+    destinyVersion: account.destinyVersion,
   };
   applyUpdateLocally(draft, updateAction);
   draft.updateQueue.push(updateAction);
 }
 
 function deleteSearch(
+  account: DestinyAccount,
   draft: Draft<DimApiState>,
-  destinyVersion: DestinyVersion,
   query: string,
   type: SearchType,
 ) {
@@ -1236,7 +1236,8 @@ function deleteSearch(
       query,
       type,
     },
-    destinyVersion,
+    platformMembershipId: account.membershipId,
+    destinyVersion: account.destinyVersion,
   };
   applyUpdateLocally(draft, updateAction);
   draft.updateQueue.push(updateAction);
@@ -1263,7 +1264,7 @@ function cleanupInvalidSearches(draft: Draft<DimApiState>, account: DestinyAccou
       customStats: draft.settings.customStats ?? [],
     } as FilterContext);
     if (!saveInHistory) {
-      deleteSearch(draft, account.destinyVersion, search.query, search.type);
+      deleteSearch(account, draft, search.query, search.type);
     }
   }
 }