+++ src/data/d2/d2-event-info.ts
@@ -12,7 +12,7 @@ export const D2EventInfo = {
     name: 'The Dawning',
     shortname: 'dawning',
     sources: [464727567, 547767158, 629617846, 2364515524, 3092212681, 3952847349, 4054646289],
-    engram: [1170720694, 2648089539, 3151770741, 3488374611],
+    engram: [1170720694, 3151770741, 3488374611],
   },
   2: {
     name: 'Crimson Days',
@@ -29,8 +29,8 @@ export const D2EventInfo = {
   4: {
     name: 'Festival of the Lost',
     shortname: 'fotl',
-    sources: [1054169368, 1677921161, 1919933822, 3190938946, 3693722471, 4041583267],
-    engram: [1451959506],
+    sources: [1054169368, 1677921161, 1919933822, 3190938946, 3482766024, 3693722471, 4041583267],
+    engram: [1451959506, 1553479004],
   },
   5: {
     name: 'The Revelry',
@@ -75,6 +75,7 @@ export const D2SourcesToEvent = {
   1677921161: D2EventEnum.FESTIVAL_OF_THE_LOST,
   1919933822: D2EventEnum.FESTIVAL_OF_THE_LOST,
   3190938946: D2EventEnum.FESTIVAL_OF_THE_LOST,
+  3482766024: D2EventEnum.FESTIVAL_OF_THE_LOST,
   3693722471: D2EventEnum.FESTIVAL_OF_THE_LOST,
   4041583267: D2EventEnum.FESTIVAL_OF_THE_LOST,
   2187511136: D2EventEnum.REVELRY,